unsigned PortNumber
PortNumber class update
- internal representation now uses int
- toString() now converts the value as unsigned integer
- renamed value() to shortValue()
- added value() method returning port number as unsigned long
- added constructor from String
- JSON serialization now uses the new value(), so that special handling for
  unsigned integer is not required

Update classes using PortNumber
- Replaced value() calls with new value(), when the return value was
  converted to long immediately.
- toString() result may change if the object contained PortNumber.
  - e.g., FlowEntryAction, FlowEntry, SwitchPort, ...

+++ src/main/java/net/onrc/onos/core/datastore/topology/KVLink.java
@@ -69,7 +69,7 @@ public class KVLink extends KVObject {
          */
         public SwitchPort(final Dpid srcDpid, final PortNumber srcPortNo) {
             this.dpid = srcDpid.value();
-            this.number = (long) srcPortNo.value();
+            this.number = srcPortNo.value();
         }
 
         /**
